chore: first pass at a list view
