+++ __mocks__/@grafana/ui.tsx
@@ -1,7 +1,7 @@
 import * as ui from '@grafana/ui';
 import React from 'react';
 
-const Select = ({ options, value, onChange, multiple = false }: any) => {
+const Select = ({ options, value, onChange, multiple = false, ...rest }: any) => {
   function handleChange(event) {
     const option = options.find((option) => {
       return String(option.value) === event.currentTarget.value;
@@ -14,7 +14,7 @@ const Select = ({ options, value, onChange, multiple = false }: any) => {
   }
 
   return (
-    <select data-testid="select" value={value} onChange={handleChange} multiple={multiple}>
+    <select data-testid={rest['data-testid'] ?? 'select'} value={value} onChange={handleChange} multiple={multiple}>
       {options.map(({ label, value }) => (
         <option key={value} value={value}>
           {label}
