+++ src/components/CheckListItem/index.ts
@@ -1,0 +1,1 @@
+export { CheckListItem } from './CheckListItem';
